Initial separation of the single h file into .c's
* Creates a static library to link against
* Moves arq_init into src/arq.c as a proof of concept
* Adds default options for previously required compile time flags
