+++ functional_tests/CMakeLists.txt
@@ -8,8 +8,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED on)
 
 include_directories(${CPPUTEST_INCLUDE_DIR})
 
-set(ARQ_FUNCTIONAL_TEST_SOURCES ${CMAKE_SOURCE_DIR}/arq.h
-                                main.cpp
+set(ARQ_FUNCTIONAL_TEST_SOURCES main.cpp
                                 functional_tests.h
                                 arq_in_functional_tests.h
                                 arq_in_functional_tests.c
@@ -44,7 +43,7 @@ set_source_files_properties(arq_in_test_project.c PROPERTIES COMPILE_FLAGS "${AR
 add_executable(arq_functional_tests ${ARQ_FUNCTIONAL_TEST_SOURCES})
 target_compile_options(arq_functional_tests PRIVATE ${ARQ_COMMON_FLAGS})
 add_dependencies(arq_functional_tests CppUTest_external)
-target_link_libraries(arq_functional_tests libCppUTest libCppUTestExt)
+target_link_libraries(arq_functional_tests libCppUTest libCppUTestExt nanoarq)
 
 if (MSVC)
     target_link_libraries(arq_functional_tests winmm)
